pyproject.toml
[tool.poetry]
name = "aml-admin"
version = "2.0.1"
description = "AML ERP System & Planners Portal"
readme = "readme.md"
homepage = "https://amlglobal.net/"
repository = "https://github.com/amlglobal/aml-admin"
authors = [
    "Denis Verbin <denis.verbin@aviationdata.aero>",
    "Lukasz Przybyla <lukasz.przybyla@aviationdata.aero>",
    "Kristof Nemet <kristof.nemet@aviationdata.aero>",
]

[tool.poetry.dependencies]
python = "^3.11"
Django = "^4.2.6"

# System packages
sentry-sdk = "^1.35.0"
python-decouple = "^3.8"
mysqlclient = "^2.2.0"
django-redis = "^5.4.0"
django-two-factor-auth = "^1.15.5"
django-database-prefix = "^0.1.0"
PyYAML = "^6.0.1"
celery = "^5.3.5"
channels = "^4.0.0"
attrs = "^23.1.0"
django-cors-headers = "^4.3.0"
daphne = "^4.0.0"
channels-redis = "^4.1.0"

# REST Framework
djangorestframework = "^3.14.0"
djangorestframework-jsonapi = "^6.1.0"
django-filter = "^23.3"


# UI Packages
django-bootstrap-modal-forms = "^3.0.4"
django-select2 = "^8.1.2"
django-ajax-datatable = "^4.4.5"
django-downloadview = "^2.3.0"
django-flatpickr = "^2.0.1"
amlopsvueelements = ">=1.6,<1.7"


# Notifications / external APIs
django-storages = "^1.14.2"
boto3 = "^1.28.85"
fcm-django = "^2.0.0"
twilio = "^8.10.1"
py-firebase-dynamic-links = "^1.0.1"
django-notifications-hq = "^1.8.3"
logtail-python = "^0.2.7"


# PDF/CSV Tools
weasyprint = "^60.1"
cssselect2 = "^0.7.0"
pypdf = "^3.16.4"
tablib = "^3.5.0"


# Web scrapping tools
selenium = "^4.14.0"
webdriver-manager = "^4.0.1"
selenium-wire = "^5.1.0"


# Other tools
django-recaptcha = "^3.0.0"
django-solo = "^2.1.0"
django-tagify2 = "^1.0.19"
django-sql-utils = "^0.7.0"
django-shortuuidfield = "^0.1.3"
phonenumberslite = "^8.13.25"
psycopg-binary = "^3.1.12"
psycopg = "^3.1.12"
grpcio = "^1.59.2"
timezonefinder = "^6.2.0"
django-cacheops = "^7.0.2"
chardet = "^5.2.0"
thefuzz = "^0.20.0"
holidays = "^0.38"


[tool.poetry.group.dev.dependencies]
django-debug-toolbar = "^4.0.0"
autopep8 = "^2.0.4"
ipython = "^8.15.0"
django-extensions = "^3.2.3"
jupyter = "^1.0.0"
notebook = "6.5.5"
traitlets = "5.9.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
